requirements.txt
gradio
pydantic
pandas
nordigen
jupyter
numpy
pydotenv
fastapi
python-dotenv
loguru
aiohttp